split merge validation service out from merge service
