+++ hapi-fhir-docs/src/main/resources/ca/uhn/hapi/fhir/docs/validation/examples.md
@@ -39,5 +39,5 @@ FhirValidator validator = myFhirCtx.newValidator();
 validator.registerValidatorModule(instanceValidator);
 
 // Validate theResource
-ValidationResult validationResult = validator.validateWithResult(theResource);
+ValidationResult mergeValidationResult = validator.validateWithResult(theResource);
 ```
